+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added a GoLand extension for running and debugging individual Testo suite tests from editor gutter actions.
+
 ## [1.6.0] - 2026-07-17
 
 ### Added